requirements.txt
# API
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
pytesseract>=0.3.8

# ML & RAG
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utils
pyyaml>=5.4.1
python-dotenv>=0.19.0
joblib>=1.0.1

# Testing
pytest>=6.2.5
pytest-asyncio>=0.18.0
httpx>=0.23.0